third_party/blink/web_tests/fast/forms/label/labelable-elements-expected.txt

Test labelable elements have labels attribute and others do not.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Labelable: button [object HTMLButtonElement]
PASS element.labels.length is 1

Not labelable: datalist [object HTMLDataListElement]
PASS element.labels is undefined.
PASS element.labels is not null

Not labelable: fieldset [object HTMLFieldSetElement]
PASS element.labels is undefined.
PASS element.labels is not null

Labelable: input [object HTMLInputElement]
PASS element.labels.length is 1

Not labelable: label [object HTMLLabelElement]
PASS element.labels is undefined.
PASS element.labels is not null

Not labelable: legend [object HTMLLegendElement]
PASS element.labels is undefined.
PASS element.labels is not null

Labelable: meter [object HTMLMeterElement]
PASS element.labels.length is 1

Not labelable: object [object HTMLObjectElement]
PASS element.labels is undefined.
PASS element.labels is not null

Not labelable: optgroup [object HTMLOptGroupElement]
PASS element.labels is undefined.
PASS element.labels is not null

Not labelable: option [object HTMLOptionElement]
PASS element.labels is undefined.
PASS element.labels is not null

Labelable: output [object HTMLOutputElement]
PASS element.labels.length is 1

Labelable: progress [object HTMLProgressElement]
PASS element.labels.length is 1

Labelable: select [object HTMLSelectElement]
PASS element.labels.length is 1

Labelable: textarea [object HTMLTextAreaElement]
PASS element.labels.length is 1

Labelable: buttonType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: checkboxType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: colorType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: dateType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: datetimeType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: datetime-localType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: emailType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: fileType [object HTMLInputElement]
PASS element.labels.length is 1

Not labelable: hiddenType [object HTMLInputElement]
PASS element.labels is null

Labelable: imageType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: monthType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: numberType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: passwordType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: radioType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: rangeType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: resetType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: searchType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: submitType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: telType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: textType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: timeType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: urlType [object HTMLInputElement]
PASS element.labels.length is 1

Labelable: weekType [object HTMLInputElement]
PASS element.labels.length is 1
PASS successfullyParsed is true

TEST COMPLETE
